# dict-based `SparsePauliOp.simplify`