--- a/qiskit/__init__.py
+++ b/qiskit/__init__.py
@@ -25,8 +25,7 @@ import qiskit._accelerate
 # manually define them on import so people can directly import
 # qiskit._accelerate.* submodules and not have to rely on attribute access
 sys.modules["qiskit._accelerate.stochastic_swap"] = qiskit._accelerate.stochastic_swap
-
-sys.modules["qiskit._accelerate.array_unique"] = qiskit._accelerate.array_unique
+sys.modules["qiskit._accelerate.sparse_pauli_op"] = qiskit._accelerate.sparse_pauli_op
 
 
 # qiskit errors operator