--- a/src/lib.rs
+++ b/src/lib.rs
@@ -17,11 +17,14 @@ use pyo3::Python;
 mod edge_collections;
 mod nlayout;
 mod stochastic_swap;
+mod array_unique;
 
 use crate::stochastic_swap::PyInit_stochastic_swap;
+use crate::array_unique::PyInit_array_unique;
 
 #[pymodule]
 fn _accelerate(_py: Python<'_>, m: &PyModule) -> PyResult<()> {
     m.add_wrapped(wrap_pymodule!(stochastic_swap))?;
+    m.add_wrapped(wrap_pymodule!(array_unique))?;
     Ok(())
 }