… releasenotes/notes/add-sparsepauliop-equiv-7a8a1420117dba21.yaml