<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add a release note file using reno if the change needs to be
  documented in the release notes.
⚠️ If your pull request fixes an open issue, please link to the issue.

- [X] I have added the tests to cover my changes.
- [X] I have updated the documentation accordingly.
- [X] I have read the CONTRIBUTING document.
-->

### Summary

This PR replaces `np.unique` in `SparsePauliOp.simplify` with a dict-based operation.
~~This is faster than `np.unique` when the number of qubits is large and there are many duplicate operators.~~
~~But, if the number of qubits is small or there is no duplicate operator, this PR can be slower.~~
~~Moreover, the code gets more complicated than just applying `np.unique`.~~
~~So, I don't have a strong opinion to include this PR as Terra. I want feedback from Qiskit developers.~~

Update:
This PR contains three updates:

1. I implemented a utility function `unordered_unique` in rust to replace `np.unique` in `SparsePauliOp.simplify`.

2. I added `SparsePauliOp.equiv` to check the equivalence of two operators while `SparsePauliOp.__eq__` checks elementwise equality. Based on #7657 
E.g.,
```python
op = SparsePauliOp.from_list([("X", 1), ("Y", 1)])
op2 = SparsePauliOp.from_list([("X", 1), ("Y", 1), ("Z", 0)])
op3 = SparsePauliOp.from_list([("Y", 1), ("X", 1)])

print(op == op2)  # False
print(op == op3)  # False
print(op.equiv(op2))  # True
print(op.equiv(op3))  # True
```

3. I found a bug that `SparsePauliOp.__eq__` raises an error if two operators have different numbers of coefficients.
I fixed it and add unit tests.

```python
op = SparsePauliOp.from_list([("X", 1), ("Y", 1)])
op2 = SparsePauliOp.from_list([("X", 1), ("Y", 1), ("Z", 0)])
print(op == op2)
#  ValueError: operands could not be broadcast together with shapes (2,) (3,)
```

Closes #7657 

### Details and comments

I attach a microbenchmark. Because `op4 = op3.simplify()`, `op4` does not contain any duplicate operator.
The current approach with `np.unique` is expected to be faster in the 2nd pass because it is fast when operators are already sorted.

```python
import random
from timeit import timeit

from qiskit.quantum_info import SparsePauliOp

def bench(k, n):
    random.seed(123)
    op = SparsePauliOp.from_list([(''.join(random.choices('IXYZ', k=k)), 1) for _ in range(n)])
    op2 = SparsePauliOp.from_list([(''.join(random.choices('IXYZ', k=k)), 1) for _ in range(n)])
    op3 = op.compose(op2)
    op4 = op3.simplify()

    print(f'{k}, {n} (1st pass): {timeit(lambda: op3.simplify(), number=1)} sec')
    print(f'{k}, {n} (2nd pass): {timeit(lambda: op4.simplify(), number=1)} sec')
    print()

for i in range(8, 41, 8):
    bench(i, 1000)
```

main
```
8, 1000 (1st pass): 0.7635101599989866 sec
8, 1000 (2nd pass): 0.014977621998696122 sec

16, 1000 (1st pass): 1.1324572909979906 sec
16, 1000 (2nd pass): 0.2923723389976658 sec

24, 1000 (1st pass): 1.223762061999878 sec
24, 1000 (2nd pass): 0.3293825949986058 sec

32, 1000 (1st pass): 1.1833479929991881 sec
32, 1000 (2nd pass): 0.2818607930021244 sec

40, 1000 (1st pass): 1.3982988319985452 sec
40, 1000 (2nd pass): 0.40111826100110193 sec
```

this PR
```
8, 1000 (1st pass): 0.3265826780007046 sec
8, 1000 (2nd pass): 0.011909335997188464 sec

16, 1000 (1st pass): 0.4464764760014077 sec
16, 1000 (2nd pass): 0.29431921200011857 sec

24, 1000 (1st pass): 0.28692975499870954 sec
24, 1000 (2nd pass): 0.27714769699741737 sec

32, 1000 (1st pass): 0.2480848600025638 sec
32, 1000 (2nd pass): 0.24830579999979818 sec

40, 1000 (1st pass): 0.2818297639969387 sec
40, 1000 (2nd pass): 0.28098882400081493 sec
```