@t-imamichi 2022-02-18
Thank you for your comments. I tried my microbenchmarks with smaller qubits. I see the current code runs faster thanks to `np.packbits` I believe.

```python
import random
from timeit import timeit

from qiskit.quantum_info import SparsePauliOp

def bench(k, n):
    random.seed(123)
    op = SparsePauliOp.from_list([(''.join(random.choices('IXYZ', k=k)), 1) for _ in range(n)])
    op2 = SparsePauliOp.from_list([(''.join(random.choices('IXYZ', k=k)), 1) for _ in range(n)])
    op3 = op.compose(op2)
    op4 = op3.simplify()

    print(f'{k}, {n} (1st pass): {timeit(lambda: op3.simplify(), number=5)} sec')
    print(f'{k}, {n} (2nd pass): {timeit(lambda: op4.simplify(), number=5)} sec')
    print()

for i in range(8, 33, 8):
    bench(i, 1000)
```

main
```
8, 1000 (1st pass): 3.351907108 sec
8, 1000 (2nd pass): 0.07600556600000008 sec

16, 1000 (1st pass): 4.943881548999999 sec
16, 1000 (2nd pass): 1.383466663 sec

24, 1000 (1st pass): 5.464609770000001 sec
24, 1000 (2nd pass): 1.619675787000002 sec

32, 1000 (1st pass): 5.373269798999999 sec
32, 1000 (2nd pass): 1.440173430999998 sec
```

this PR
```
8, 1000 (1st pass): 4.607136525 sec
8, 1000 (2nd pass): 0.23311107700000022 sec

16, 1000 (1st pass): 5.703168041 sec
16, 1000 (2nd pass): 4.6930094539999985 sec

24, 1000 (1st pass): 4.611978835999999 sec
24, 1000 (2nd pass): 4.573024002 sec

32, 1000 (1st pass): 4.659426916999998 sec
32, 1000 (2nd pass): 4.521127856 sec
```