@jakelishman 2022-02-18
Hmm, that poses some interesting questions, especially because the current version is significantly better on close-to-simplified forms.  I suspect that it's to do with being close-to-sorted as well, so shuffling the order of the operators in the output of a `simplify` probably returns `main` to the "1st pass" speeds, despite already being simplified.

The speed-boost is significant at high numbers of qubits (of course, because of the asymptotic scaling removing a factor of `ln(n)`), so it may well be worth having both, and have an option to select.

That said, actually the best solution might be to move the backing of `SparsePauliOp` down into Rust or Cython, and enforce that the Paulis are _always_ sorted.  That would let us implement the fused `_add -> simplify` in linear time, because the algorithm becomes effectively a slight variant on the "merge" primitive of mergesort, since we can guarantee that the two arrays are sorted. `compose` might be slightly tricky to code up in that style, but if we could manage it, then `simplify` would become unnecessary and the `==` check could be done in guaranteed linear time as well.  Creation of `SparsePauliOp` would become `O(n * ln(n))` instead, but that may well be a price worth paying - we'd be able to add a `sort=True` keyword argument to the constructor to skip the check if the input was guaranteed in the right format, and possibly for quite a lot of uses, the initial creation of the objects from Python objects isn't the bottleneck.