@t-imamichi 2022-03-03
I tried a rust version of `unique` in `SparsePauliOp.simplify`. I'm not familiar with rust, so the code might not be optimized well. But, the rust version is faster than sort-based `np.unique`. If anyone has an idea to optimize it, feel free to let me know.

comparison with main
```python
import random
from timeit import timeit

from qiskit.quantum_info import SparsePauliOp

def bench(k, n):
    random.seed(123)
    op = SparsePauliOp.from_list([(''.join(random.choices('IXYZ', k=k)), 1) for _ in range(n)])
    op2 = SparsePauliOp.from_list([(''.join(random.choices('IXYZ', k=k)), 1) for _ in range(n)])
    op3 = op.compose(op2)
    op4 = op3.simplify()

    print(f'{k}, {n} (1st pass): {timeit(lambda: op3.simplify(), number=1)} sec')
    print(f'{k}, {n} (2nd pass): {timeit(lambda: op4.simplify(), number=1)} sec')
    print()

for i in range(8, 41, 8):
    bench(i, 1000)
```

main
```
8, 1000 (1st pass): 0.630255923 sec
8, 1000 (2nd pass): 0.01311622400000001 sec

16, 1000 (1st pass): 0.9540651409999996 sec
16, 1000 (2nd pass): 0.2669846600000003 sec

24, 1000 (1st pass): 1.052437779 sec
24, 1000 (2nd pass): 0.3130550200000002 sec

32, 1000 (1st pass): 1.0365809030000008 sec
32, 1000 (2nd pass): 0.2829057749999997 sec

40, 1000 (1st pass): 1.219256982000001 sec
40, 1000 (2nd pass): 0.4384409550000008 sec
```

this PR
```
8, 1000 (1st pass): 0.3164088220000001 sec
8, 1000 (2nd pass): 0.021032002999999966 sec

16, 1000 (1st pass): 0.890976802 sec
16, 1000 (2nd pass): 0.665050409 sec

24, 1000 (1st pass): 0.6271690339999996 sec
24, 1000 (2nd pass): 0.580206091 sec

32, 1000 (1st pass): 0.6370045559999999 sec
32, 1000 (2nd pass): 0.5719273319999996 sec

40, 1000 (1st pass): 0.6434421850000014 sec
40, 1000 (2nd pass): 0.625335218 sec
```