@t-imamichi 2022-03-03
Thank you for your feedback, @mtreinish! I will update my rust code based on your comments. I'm really impressed with the rust framework for terra. I feel it is easier than cython. Kudos to you.