@t-imamichi 2022-03-03
Thanks to @mtreinish's comment, `simplify` gets much faster.

```
8, 1000 (1st pass): 0.16840985499999994 sec
8, 1000 (2nd pass): 0.008250524000000148 sec

16, 1000 (1st pass): 0.2780222299999999 sec
16, 1000 (2nd pass): 0.12144982699999995 sec

24, 1000 (1st pass): 0.13483429299999994 sec
24, 1000 (2nd pass): 0.14022761000000017 sec

32, 1000 (1st pass): 0.1404055569999998 sec
32, 1000 (2nd pass): 0.11580450200000003 sec

40, 1000 (1st pass): 0.15199184499999951 sec
40, 1000 (2nd pass): 0.15105953399999983 sec
```