@t-imamichi 2022-03-04
Thanks. The rust version looks good, so I will finalize this PR.
This new version of `simplify` does not guaranteed that the output is sorted.
So, the current unit tests based on `__eq__` do not pass due to the order https://github.com/Qiskit/qiskit-terra/pull/7656#issuecomment-1037991612.
The current `__eq__` checks the equality of object level (order matters). E.g., `I + X != X + I`
We perhaps need to make another method, e.g., `equiv`, that checks the equality of numerical level (order ignored) https://github.com/Qiskit/qiskit-terra/issues/7657#issuecomment-1038855386. E.g., `I + X == X + I`

TODOs
- [ ] discussion #7657 
- [ ] code cleanup
- [ ] docstrings