@t-imamichi 2022-03-07
I have an issue that pylint cannot import `qiskit._accelerate.sparse_pauli_op`.
Do you have any suggestion to fix it, @mtreinish? Am I missing anything?