@mtreinish 2022-03-07
Yeah, it's a limitation with pylint, it's using it's custom ast parsing stuff and doesn't know how to deal with compiled extensions (or the manual `sys.modules` additions we do for the compiled submodules). We probably need to update the pylintrc to handle this, but in the meantime I've just been using `#pylint: disable=import-error` on the `from qiskit._accelerate.* import` line.