@t-imamichi 2022-03-14
FYI: I updated the summary of this PR.
https://github.com/Qiskit/qiskit-terra/pull/7656#issue-1135486833